Created blobDatabase which is s3Database minus aws references
